move build logic to tauri-build
