+++ core/tauri-build/Cargo.toml
@@ -25,6 +25,7 @@ cargo_toml = "0.13"
 serde_json = "1"
 heck = "0.4"
 json-patch = "0.2"
+walkdir = "2"
 
 [target."cfg(windows)".dependencies]
 winres = "0.1"
